chatbot/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
flask>=3.0.0
flask-cors>=4.0.0
langchain>=0.1.0
langchain-huggingface>=0.0.1
langchain-community>=0.1.0
langchain-core>=0.1.0
sentence-transformers>=2.2.2
transformers>=4.36.0
faiss-cpu>=1.7.4
torch>=2.0.0
pandas>=2.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
python-dotenv>=1.0.0
langchain-text-splitters>=0.0.1
ctransformers>=0.2.27
